Factor client caches into its own code
All the bits to do with client caches can go together, since (after
being constructed) it only interacts with the reconciler through
`watchTargetAndGetReader` and via events.

This is mainly a case of relocating the relevant fields and changing
some variable names.
